refactor(jolt-witness): colocate Extract impls with their newtypes
Each witnesses/ file now holds a family's complete definition — value
newtype, field encoding, and its single-sourced trace derivation. The
Extract/ExtractIndexed traits, WitnessEnv, and the shared row helpers
move to witnesses/mod.rs; the id-driven dispatcher stays backend-side
in trace.rs. Pure code motion (byte-diff 10/10).
