+++ crates/jolt-witness/src/witnesses/operands.rs
@@ -1,4 +1,13 @@
-use jolt_field::{signed::S128, Field};
+use jolt_field::{
+    signed::{S128, S64},
+    Field,
+};
+use jolt_lookup_tables::LookupQuery;
+use jolt_program::execution::TraceRow;
+
+use super::{lookup_query, Extract, WitnessEnv};
+use crate::protocols::jolt_vm::RV64_XLEN;
+use crate::WitnessError;
 
 /// Left lookup operand of the instruction's lookup query.
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
@@ -31,9 +40,31 @@ impl LeftLookupOperand {
     }
 }
 
+impl Extract for LeftLookupOperand {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        let (left, _) = LookupQuery::<RV64_XLEN>::to_lookup_operands(&lookup_query(row));
+        Ok(Self(left))
+    }
+}
+
 impl RightLookupOperand {
     pub fn to_field<F: Field>(self) -> F {
         F::from_u128(self.0)
+    }
+}
+
+impl Extract for RightLookupOperand {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        let (_, right) = LookupQuery::<RV64_XLEN>::to_lookup_operands(&lookup_query(row));
+        Ok(Self(right))
     }
 }
 
@@ -43,9 +74,31 @@ impl LeftInstructionInput {
     }
 }
 
+impl Extract for LeftInstructionInput {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        let (left, _) = LookupQuery::<RV64_XLEN>::to_instruction_inputs(&lookup_query(row));
+        Ok(Self(left))
+    }
+}
+
 impl RightInstructionInput {
     pub fn to_field<F: Field>(self) -> F {
         F::from_i128(self.0)
+    }
+}
+
+impl Extract for RightInstructionInput {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        let (_, right) = LookupQuery::<RV64_XLEN>::to_instruction_inputs(&lookup_query(row));
+        Ok(Self(right))
     }
 }
 
@@ -66,8 +119,31 @@ impl Product {
     }
 }
 
+impl Extract for Product {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        let (left, right) = LookupQuery::<RV64_XLEN>::to_instruction_inputs(&lookup_query(row));
+        Ok(Self(
+            S64::from_u64(left).mul_trunc::<2, 2>(&S128::from_i128(right)),
+        ))
+    }
+}
+
 impl Imm {
     pub fn to_field<F: Field>(self) -> F {
         F::from_i128(self.0)
+    }
+}
+
+impl Extract for Imm {
+    fn extract(
+        row: &TraceRow,
+        _next: Option<&TraceRow>,
+        _env: &WitnessEnv<'_>,
+    ) -> Result<Self, WitnessError> {
+        Ok(Self(row.instruction.operands.imm))
     }
 }
